Extract TraceSerializingDisruptor class from DDAgentWriter
